Focus window under pointer on (re)start
When CWM is started for the first time, or restarted, check to see if there
is a client under the mouse pointer's position and focus it.  This stops the
need (particularly on restart) of having to cause an EnterNotify even on a
client by moving the mouse, just to focus it.