Config loading: don't bail on errors
Currently, when CWM is parsing its file, on the first sign of trouble it
throws out the validity of the file and assumes to use its own defaults.
This seems strange, hence now we let any synatx errors go to STDERR, and
continue to parse the file as much as we can.