Init and use FIFO for status output
This introduces the ability for CWM to output information down a named FIFO
when status events happen.  Currently, the scope of this is quite limited
but it provides just enough information to be useful.

There is one FIFO per screen defined, which should make separating out
events per screen easier.