Raise/Lower fullscreen windows
At the moment when CWM receives a _NET_WM_STATE_FULLSCREEN request or a
command to fullscreen a window, will do just that.  However, it's arguable
that the window should be raised when entering fullscreen, and lowered again
when restoring the original state.

The rationale behind this is that often if there's windows on top of a web
browser, and the user chooses to fullscreen a HTML5 video from the browser,
that the fullscreened window should be on top, and put back in its original
layer when that's no longer true, so that the window is seen.