Honor _NET_WM_DESKTOP use of -1 for sticky groups
The EWMH spec says that the cardinal returned for _NET_WM_DESKTOP can be
0xFFFFFF or -1, which should indicate that the window appears on all
desktops.  So for now, put such windows in group 0.

Without this, the range checking in group_movetogroup() was bailing cwm in
error, because -1 is quite correctly not in range of the groups we declare.

This should fix reports of stalonetray "crashing" CWM, etc.