groups: group_cycle to include all groups
Now that groups are per-monitor, when cycling through groups, ensure ALL
groups are allowed to be switched to, and not just those which contain
clients.

This is useful to allow one to switch to the group a window will be on,
rather than having to assign it to a group after it has been mapped.