client_init: fix recapture/moving of windows
When clients were previously recaptured on restart of cwm, the windows would
be moved relative to the border size.  However, this would have resulted in
windows moving relative to a border size which was set based on the config,
not the window's actual size.

Therefore, calculate this based on what the window attributes are, and then
add back the configured borderwidth afterwards and refresh the window.