client_placecalc: don't assume GLOBAL_MONITOR
When working out the position of a client which has been mapped, we get the
coordinates of the global monitor.

The use of GLOBAL_MONITOR to mean the root window across displays, isn't
available on RandR outputs.  As such, if cwm-new requests this, the monitor
cannot be found and cwm-new aborts.

Instead, use the Display{Width,Height} macros.