status: fix memory leak on group_init
When initialising groups, group 0 is set as the active group.  Doing this will
cause status to be output.  This is ineffective, as it will also be in the
process of creating moe groups, leading to memory corruption in the JSON
output.

Instead, output the status information as each group is generated.